+++ CHANGELOG.md
@@ -16,6 +16,7 @@ From version 2.6.0, the sections in this file adhere to the [keep a changelog](h
 
 ### Added
 * [#2190](https://github.com/Shopify/shopify-cli/pull/2190): Better login experience with spinner
+* [#2222](https://github.com/Shopify/shopify-cli/pull/2222): Adds support for `can_access_network` config for checkout extensions
 
 ## Version 2.15.2 - 2022-03-28